completed interface lift
