+++ kmc-sa-mgmt/kmc-sa-cli/src/main/java/gov/nasa/jpl/ammos/asec/kmc/cli/crud/SaCreate.java
@@ -1,6 +1,7 @@
 package gov.nasa.jpl.ammos.asec.kmc.cli.crud;
 
 import gov.nasa.jpl.ammos.asec.kmc.api.ex.KmcException;
+import gov.nasa.jpl.ammos.asec.kmc.api.sa.ISecAssn;
 import gov.nasa.jpl.ammos.asec.kmc.api.sa.SecAssn;
 import gov.nasa.jpl.ammos.asec.kmc.api.sa.SecAssnValidator;
 import gov.nasa.jpl.ammos.asec.kmc.api.sa.SpiScid;
@@ -20,7 +21,6 @@ import java.util.List;
 
 /**
  * Create a Security Assocation
- *
  */
 @CommandLine.Command(name = "create", description = "Create a new Security Association", mixinStandardHelpOptions =
         true, versionProvider = Version.class)
@@ -84,8 +84,8 @@ public class SaCreate extends BaseCreateUpdate {
     protected void doSingle() throws KmcException {
         console(String.format("%s creating SA", user));
         try (IKmcDao dao = getDao()) {
-            SpiScid id = new SpiScid(spi, scid);
-            SecAssn sa = dao.getSa(id);
+            SpiScid  id = new SpiScid(spi, scid);
+            ISecAssn sa = dao.getSa(id);
             if (sa != null) {
                 throwEx(String.format("Create error, SA %d/%d already exists", id.getSpi(), id.getScid()));
             }
@@ -136,7 +136,8 @@ public class SaCreate extends BaseCreateUpdate {
                         checkEncParams(args.single.optionalArgs.ekid, args.single.optionalArgs.ecs);
                         checkAuthParams(args.single.optionalArgs.akid, args.single.optionalArgs.acs);
                         checkSt(args.single.optionalArgs.st);
-                        checkIvParams(args.single.optionalArgs.iv, args.single.optionalArgs.ivLen, args.single.optionalArgs.st, args.single.optionalArgs.ecs );
+                        checkIvParams(args.single.optionalArgs.iv, args.single.optionalArgs.ivLen,
+                                args.single.optionalArgs.st, args.single.optionalArgs.ecs);
                         checkArsnParams(args.single.optionalArgs.arsn, args.single.optionalArgs.arsnlen);
                         checkArsnWParams(args.single.optionalArgs.arsnw);
                         checkAbmParams(args.single.optionalArgs.abm, args.single.optionalArgs.abmLen);
