+++ kmc-sa-mgmt/kmc-sa-gui/src/main/frontend/src/setupProxy.js
@@ -4,7 +4,7 @@ module.exports = function(app) {
     app.use(
         '/api',
         createProxyMiddleware({
-            target: 'http://localhost:8080',
+            target: 'http://localhost:8447',
             changeOrigin: true,
         })
     );
