+++ kmc-sa-mgmt/kmc-sa-gui/src/main/resources/etc/kmc-sa-mgmt-dev.properties
@@ -1,0 +1,39 @@
+#
+# Copyright 2022, by the California Institute of Technology.
+# ALL RIGHTS RESERVED. United States Government Sponsorship acknowledged.
+# Any commercial use must be negotiated with the Office of Technology
+# Transfer at the California Institute of Technology.
+#
+# This software may be subject to U.S. export control laws. By accepting
+# this software, the user agrees to comply with all applicable U.S.
+# export laws and regulations. User has the responsibility to obtain
+# export licenses, or other export authority as may be required before
+# exporting such information to foreign countries or providing access to
+# foreign persons.
+#
+# database user name
+db.auth.user=sadb_user
+# database password
+db.auth.pass=sadb_password
+# database connection string. if provided, the db.host, db.port, and db.schema keys will be ignored
+db.conn.string=none
+# database host
+db.host=localhost
+# database port
+db.port=3307
+# database schema
+db.schema=sadb
+# database TLS keystore
+db.keystore=none
+# database TLS keystore password
+db.keystore.pass=none
+# database TLS truststore
+db.truststore=none
+# database TLS truststore password
+db.truststore.pass=none
+# use TLS for database connections
+# the provided truststore must contain the certificate chain for connecting to the database.
+db.tls=false
+# use mutual TLS for database connections. when mTLS is true, the db.auth.pass key is ignored.
+# the provided keystore must be trusted by the database server and database user.
+db.mtls=false
